feat: templatize portal
